pyproject.toml
[tool.poetry]
name = "sprint_4"
version = "0.1.0"
description = ""
authors = ["DmitryK <kachkindmitry@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8.1"
elasticsearch = "^7.17.1"
fastapi = "^0.75.0"
aioredis = "^2.0.1"
orjson = "^3.6.7"
elasticsearch-dsl = "^7.4.0"
uvicorn = "^0.17.6"
python-dotenv = "^0.19.2"
fastapi-cache2 = {extras = ["redis"], version = "^0.1.8"}
backoff = "^1.11.1"
aiohttp = "^3.8.1"
redis = "^4.1.4"
PyJWT = "^2.3.0"



[tool.poetry.dev-dependencies]
elasticsearch = "^7.17.1"
fastapi = "^0.75.0"
aioredis = "^2.0.1"
orjson = "^3.6.7"
elasticsearch-dsl = "^7.4.0"
fastapi-cache2 = {extras = ["redis"], version = "^0.1.8"}
uvicorn = "^0.17.6"
python-dotenv = "^0.19.2"
backoff = "^1.11.1"
psycopg2-binary = "^2.9.3"
schedule = "^1.1.0"
aiohttp = "^3.8.1"
aiofiles = "^0.8.0"
pytest = "^7.1.1"
pytest-asyncio = "^0.18.2"
PyJWT = "^2.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode= "strict"
log_cli = true
log_cli_level = "INFO"
log_cli_format = "%(levelname)-5.5s [%(name)s] %(message)s"